o The developing branch 'branch_pqueue' is merged   back to the main trunk. o No further development is done on 'branch_pqueue',   the cvs main trunk is now the current development branch. o You need to port your client code, see xmlBlaster/CHANGES for details
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@7326 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/java/org/xmlBlaster/engine/admin/I_AdminSession.java
@@ -22,9 +22,9 @@ public interface I_AdminSession {
    /** How many messages where sent to this clients login session */
    public long getNumUpdates();
    /** How many messages are in this clients session callback queue */
-   public int getCbQueueNumMsgs();
+   public long getCbQueueNumMsgs();
    /** How many messages are max. allowed in this clients session callback queue */
-   public int getCbQueueMaxMsgs();
+   public long getCbQueueMaxMsgs();
    /** Destroy the session (force logout) */
    public String getKillSession() throws XmlBlasterException;
 }
